mozilla-release/dom/push/test/test_has_permissions.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1038811: Push tests.

Any copyright is dedicated to the Public Domain.
http://creativecommons.org/licenses/publicdomain/

-->
<head>
  <title>Test for Bug 1038811</title>
<<<<<<< HEAD
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/AddTask.js"></script>
||||||| merged common ancestors
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
=======
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
>>>>>>> upstream-releases
  <script type="text/javascript" src="/tests/dom/push/test/test_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8">
</head>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1038811">Mozilla Bug 1038811</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="text/javascript">
  /* import-globals-from ./test_utils.js */

  function debug(str) {
  // console.log(str + "\n");
  }

<<<<<<< HEAD
  add_task(async function start() {
    await setupPrefsAndMockSocket(new MockWebSocket());
||||||| merged common ancestors
  function start() {
    return navigator.serviceWorker.register("worker.js" + "?" + (Math.random()), {scope: "."})
    .then((swr) => {
      registration = swr;
      return waitForActive(registration);
    });
  }
=======
  var registration;
>>>>>>> upstream-releases

<<<<<<< HEAD
    var url = "worker.js" + "?" + Math.random();
    registration = await navigator.serviceWorker.register(url, {scope: "."});
    await waitForActive(registration);
  });
||||||| merged common ancestors
  function unregister() {
    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    }, function(e) {
      dump("Unregistering the SW failed with " + e + "\n");
    });
  }
=======
  add_task(async function start() {
    await setupPrefsAndMockSocket(new MockWebSocket());
>>>>>>> upstream-releases

<<<<<<< HEAD
  add_task(async function hasPermission() {
    var state = await registration.pushManager.permissionState();
    debug("state: " + state);
    ok(["granted", "denied", "prompt"].includes(state), "permissionState() returned a valid state.");
  });
||||||| merged common ancestors
  function hasPermission(swr) {
    var p = new Promise(function(res, rej) {
      swr.pushManager.permissionState().then(
        function(state) {
          debug("state: " + state);
          ok(["granted", "denied", "prompt"].includes(state), "permissionState() returned a valid state.");
          res(swr);
        }, function(error) {
          ok(false, "permissionState() failed.");
          res(swr);
        }
        );
    });
    return p;
  }

  function runTest() {
    start()
    .then(hasPermission)
    .then(unregister)
    .catch(function(e) {
      ok(false, "Some test failed with error " + e);
    }).then(SimpleTest.finish);
  }
=======
    var url = "worker.js?" + Math.random();
    registration = await navigator.serviceWorker.register(url, {scope: "."});
    await waitForActive(registration);
  });

  add_task(async function hasPermission() {
    var state = await registration.pushManager.permissionState();
    debug("state: " + state);
    ok(["granted", "denied", "prompt"].includes(state), "permissionState() returned a valid state.");
  });
>>>>>>> upstream-releases

  add_task(async function unregister() {
    var result = await registration.unregister();
    ok(result, "Unregister should return true.");
  });

</script>
</body>
</html>
